Move ManagerForm.cs
